producers/requirements.txt
kafka-python==2.0.2
PyYAML==6.0.1
websocket-client==1.6.2
